Fix deprecated "jsonSerialize()" return type in PHP 8.1
In PHP 8.1 the return type of "jsonSerialize()" was set to "mixed", and
a "deprecated" notice is raised if subclasses do not declare any return
type. "mixed" was added in PHP 8.0, but the return type of a method can
have a more specific type in a subclass than in the parent class
(https://wiki.php.net/rfc/mixed_type_v2), so "array" can be used to
solve the deprecated notice while also being compatible with PHP 7.4.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>